OpenAI-Demo-Projects/Telegram Chatbot Project/requirements.txt
openai
python-dotenv
aiogram